fix: posts order
It seems that there was a pretty awful filter there, calling collection.reverse(). The thing is - it modifies the original collection in place. From this point on, every page calling it again, was causing a complete chaos.

Filters updated, sorting and ordering - updated as well. Now it seems working properly.