fix(ui): updating glitch
Glitch animation was causing issues, when a link was multi-lined. The reason is the display: inline-block prop. It caused the first line to be moved where the second is. If you were hovering on that first part, it caused an indefinite flickering, moving pieces up and down (a bad type of glitch :))

I'd assume the inline-block was needed because having it with bold was causing other types of issues. But since the font used is monospaced, everything stays in place.

Anyway, it fixes the initial issue and I don't see any side effects.

Fixes #22